Use FBX files for animations
Add Blender python scripts for processing .fbx files:
    - extract_hierarchy.py: Generates bone hierarchy from FBX
    - extract_rest_pose.py: Extracts rest pose from FBX
    - fbx_to_smpl.py: Converts FBX animations to JSON
    - sync_smpl.sh: script to regenerate all animation assets
    - Store source FBX files in fbx/ directory (via Git LFS)

Minified JSON keys (pelvis → p, keyframes → kf, time → t, etc.)
Reduced quaternion precision (6→4 decimal places).
Omit identity rotations and zero root positions.

Remove per-exercise placeholder JSON files.

+++ wasm/src/bone/cache.rs
@@ -4,15 +4,15 @@ use glam::{Quat, Vec3A};
 /// Dirty flags for lazy forward kinematics evaluation.
 /// Uses a bitset where bit i corresponds to BoneId with index i.
 #[derive(Debug, Clone, Copy, Default)]
-pub struct DirtyFlags(u16);
+pub struct DirtyFlags(u32);
 
-const fn compute_descendant_masks() -> [u16; BoneId::COUNT] {
-    let mut masks = [0u16; BoneId::COUNT];
+const fn compute_descendant_masks() -> [u32; BoneId::COUNT] {
+    let mut masks = [0u32; BoneId::COUNT];
     let mut i = 0;
     while i < BoneId::COUNT {
         // We are computing mask for bone 'i' (the ancestor)
         // Check every other bone 'j' to see if it is a descendant of 'i'
-        let mut mask: u16 = 0;
+        let mut mask: u32 = 0;
         let mut j = 0;
 
         while j < BoneId::COUNT {
@@ -50,7 +50,7 @@ const fn compute_descendant_masks() -> [u16; BoneId::COUNT] {
     masks
 }
 
-const DESCENDANT_MASKS: [u16; BoneId::COUNT] = compute_descendant_masks();
+const DESCENDANT_MASKS: [u32; BoneId::COUNT] = compute_descendant_masks();
 
 impl DirtyFlags {
     /// Create with all bones marked dirty
